requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
bcrypt
PyJWT
pydantic
httpx
rich
typer
python-dotenv
python-multipart